DRY for integration test setup
