+++ packages/selenium/test/RunPartial/IncludeExcludeTests.cs
@@ -7,32 +7,38 @@ namespace Deque.AxeCore.Selenium.Test.RunPartial
     public class IncludeExcludeTests : TestBase
     {
         [Test]
-        public void ShouldNotThrowWithIncludeExclude()
+        [TestCase("Chrome")]
+        public void ShouldNotThrowWithIncludeExclude(string browser)
         {
+            InitDriver(browser);
             GoToResource("context.html");
 
-            var axe = new AxeBuilder(driver)
+            var axe = new AxeBuilder(WebDriver)
                 .Include(".include")
                 .Exclude(".exclude");
             Assert.DoesNotThrow(() => axe.Analyze());
         }
 
         [Test]
-        public void ShouldNotThrowWithInclude()
+        [TestCase("Chrome")]
+        public void ShouldNotThrowWithInclude(string browser)
         {
+            InitDriver(browser);
             GoToResource("context.html");
 
-            var axe = new AxeBuilder(driver)
+            var axe = new AxeBuilder(WebDriver)
                 .Include(".include");
             Assert.DoesNotThrow(() => axe.Analyze());
         }
 
         [Test]
-        public void ShouldNotThrowWithExclude()
+        [TestCase("Chrome")]
+        public void ShouldNotThrowWithExclude(string browser)
         {
+            InitDriver(browser);
             GoToResource("context.html");
 
-            var axe = new AxeBuilder(driver)
+            var axe = new AxeBuilder(WebDriver)
                 .Exclude(".exclude");
             Assert.DoesNotThrow(() => axe.Analyze());
         }
